requirements.txt
# 经过修复和优化的 Python 依赖包列表
# 使用 "pip install -r requirements_fixed.txt" 命令进行安装

# 用于读取和处理 .csv 文件
pandas

# 用于发起 HTTP 请求，调用 Hugging Face Inference API
requests

# 用于显示优雅的进度条
tqdm

# 用于从 .env 文件中加载环境变量 (如 Hugging Face Token)
python-dotenv

# Hugging Face 官方库，用于与 Hub 交互
huggingface-hub

openai

tenacity